+++ client/src/commonMain/kotlin/org/apptank/horus/client/control/helper/IOperationDatabaseHelper.kt
@@ -99,12 +99,12 @@ interface IOperationDatabaseHelper {
     fun queryRecords(builder: QueryBuilder): List<DataMap>
 
     /**
-     * Executes a query using the provided SimpleQueryBuilder and returns the count of records.
-     * @param builder the SimpleQueryBuilder used to build the SQL query.
+     * Executes a query using the provided QueryBuilder and returns the count of records.
+     * @param builder the QueryBuilder used to build the SQL query.
      *
      * @return the count of records from the query result.
      */
-    fun countRecords(builder: SimpleQueryBuilder): Int
+    fun countRecords(builder: QueryBuilder): Int
 
 
     /**