+++ client/src/commonMain/kotlin/org/apptank/horus/client/database/builder/JoinableQueryBuilder.kt
@@ -47,7 +47,7 @@ class JoinableQueryBuilder(
      *
      * @return The current instance of [JoinableQueryBuilder] for method chaining.
      */
-    fun selectCount(): JoinableQueryBuilder {
+    override fun selectCount(): JoinableQueryBuilder {
         selectCount = true
         return this
     }